nyayasetu/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
groq
python-dotenv
httpx
